requirements.txt
fastapi
uvicorn
streamlit
requests
openai
tavily-python
